Mark added that he is PRESENT
I was trying to do it through command line and cloned the file to my master, made a new branch and added the file to that branch. Then, I edited it there. Saved it and then added it to so I essentially tracked its changed. Then, did the final merge of that branch where I edited it to the Master. Then I try git push and it gives me this error: "fatal: unable to access 'https://github.com/reedacartwright/Genomic-Analysis/': The requested URL returned error: 403". And even when I go to the Genomic Analysis repository and try to submit a pull request, I seem to be coming up empty.

So this is why I am attempting to do it through here.